+++ R/lm.R
@@ -80,22 +80,7 @@ train_tslm <- function(.data, specials, ...) {
 
 specials_tslm <- new_specials(
   common_xregs,
-  xreg = function(...) {
-    model_formula <- new_formula(
-      lhs = NULL,
-      rhs = reduce(enexprs(...), function(.x, .y) call2("+", .x, .y))
-    )
-    env <- parent.frame()
-    if (!exists("list", env)) env <- base_env()
-
-    env$lag <- lag # Mask user defined lag to retain history when forecasting
-    xreg <- model.frame(model_formula, data = env, na.action = stats::na.pass)
-    mm <- model.matrix(terms(xreg), xreg)
-    if (NROW(mm) == 0 && identical(colnames(mm), "(Intercept)")) {
-      return(matrix(data = 1, nrow = NROW(self$data), dimnames = list(NULL, "(Intercept)")))
-    }
-    mm
-  },
+  xreg = special_xreg(),
   .required_specials = "xreg",
   .xreg_specials = names(common_xregs),
 )
@@ -115,7 +100,7 @@ specials_tslm <- new_specials(
 #' @section Specials:
 #'
 #' \subsection{xreg}{
-#' Exogenous regressors can be included in an ARIMA model without explicitly using the `xreg()` special. Common exogenous regressor specials as specified in [`common_xregs`] can also be used. These regressors are handled using [stats::model.frame()], and so interactions and other functionality behaves similarly to [stats::lm()].
+#' Exogenous regressors can be included in a TSLM model without explicitly using the `xreg()` special. Common exogenous regressor specials as specified in [`common_xregs`] can also be used. These regressors are handled using [stats::model.frame()], and so interactions and other functionality behaves similarly to [stats::lm()].
 #' \preformatted{
 #' xreg(...)
 #' }
@@ -262,7 +247,14 @@ report.TSLM <- function(object, digits = max(3, getOption("digits") - 3), ...) {
 }
 
 #' @inherit forecast.ARIMA
+#' 
 #' @importFrom stats predict
+#'
+#' @param approx_normal Should the resulting forecast distributions be
+#'   approximated as a Normal distribution instead of a Student's T 
+#'   distribution. Returning Normal distributions (the default) is a useful 
+#'   approximation to make it easier for using TSLM models in model combinations 
+#'   or reconciliation processes.
 #'
 #' @examples
 #' as_tsibble(USAccDeaths) %>%
@@ -270,7 +262,7 @@ report.TSLM <- function(object, digits = max(3, getOption("digits") - 3), ...) {
 #'   forecast()
 #' @export
 forecast.TSLM <- function(object, new_data, specials = NULL, bootstrap = FALSE,
-                          times = 5000, ...) {
+                          approx_normal = TRUE, times = 5000, ...) {
   coef <- object$coefficients
   rank <- object$rank
   qr <- object$qr
@@ -292,7 +284,7 @@ forecast.TSLM <- function(object, new_data, specials = NULL, bootstrap = FALSE,
       map(as.numeric)
     distributional::dist_sample(sim)
   } else {
-    fc <- xreg[, piv, drop = FALSE] %*% coef[piv]
+    fc <- drop(xreg[, piv, drop = FALSE] %*% coef[piv])
     resvar <- object$sigma2
 
     if (rank > 0) {
@@ -303,8 +295,12 @@ forecast.TSLM <- function(object, new_data, specials = NULL, bootstrap = FALSE,
       ip <- rep(0, length(fc))
     }
 
-    se <- sqrt(ip + resvar)
-    distributional::dist_normal(drop(fc), drop(se))
+    se <- drop(sqrt(ip + resvar))
+    if(approx_normal){
+      distributional::dist_normal(fc, se)
+    } else {
+      distributional::dist_student_t(object$df.residual, fc, se)
+    }
   }
 }