+++ tests/testthat/test-rw.R
@@ -36,14 +36,9 @@ test_that("RW w/ drift", {
     unclass(forecast_fc$mean)
   )
 
-  if(packageVersion("fabletools") > "0.1.3"){
+  if(packageVersion("forecast") > "8.17.0"){
     expect_equivalent(
       unclass(hilo(fable_fc)$`80%`)$upper,
-      unclass(forecast_fc$upper[, 1])
-    )
-  } else {
-    expect_equivalent(
-      hilo(fable_fc)$`80%`$.upper,
       unclass(forecast_fc$upper[, 1])
     )
   }